shmoke/requirements.txt
aiocache==0.11.1
aiofile==3.7.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiologger==0.6.1
aiopg==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
blacksheep==1.0.9
blis==0.7.4
bs4==0.0.1
cached-property==1.5.2
caio==0.9.3
catalogue==1.0.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cymem==2.0.5
Discord-Anti-Spam==0.9.0
discord.py==1.7.3
Django==3.2.6
dnspython==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
essentials==1.1.4
essentials-openapi==0.1.5
fuzzywuzzy==0.18.0
gevent==21.1.2
greenlet==1.1.0
guardpost==0.0.7
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==1.1.0
Jinja2==3.0.1
keyring==23.0.1
mariadb==1.0.7
MarkupSafe==2.0.1
more-itertools==8.8.0
multidict==5.1.0
murmurhash==1.0.5
numpy==1.21.1
oauthlib==3.1.1
ordered-set==3.1.1
ordered-set-stubs==0.1.3
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.5
profanity-filter==1.3.3
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
PyNaCl==1.4.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.0
python-Levenshtein==0.12.2
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rodi==1.1.1
ruamel.yaml==0.15.100
six==1.16.0
sonyflake-py==1.3.0
soupsieve==2.2.1
spacy==2.3.7
SQLAlchemy==1.4.23
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
tomlkit==0.5.11
tqdm==4.62.0
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
wasabi==0.8.2
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
